Fix ambiguous amount multiplication
Summary: Similar to D11362.

Test Plan:
On OSX:
  ninja all check-avalanche

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Differential Revision: https://reviews.bitcoinabc.org/D11374

+++ src/avalanche/test/init_tests.cpp
@@ -19,7 +19,8 @@ BOOST_FIXTURE_TEST_SUITE(init_tests, BasicTestingSetup)
 
 BOOST_AUTO_TEST_CASE(avalanche_flag_tests) {
     gArgs.ForceSetArg("-ecash", "1");
-    const Amount tenBillion = 10'000'000'000 * Currency::get().baseunit;
+    const Amount tenBillion =
+        int64_t(10'000'000'000) * Currency::get().baseunit;
 
     auto getAvaMinQuorumStakeAmount = [](const ArgsManager &args,
                                          const Amount defaultAmount) {
