+++ src/init.cpp
@@ -1760,8 +1760,9 @@ void InitParameterInteraction(ArgsManager &args) {
         args.SoftSetBoolArg("-automaticunparking", !fAvalanche);
         args.SoftSetArg(
             "-avaminquorumstake",
-            fAvalanche ? FormatMoney(1'000'000'000'000 * SATOSHI) // 10B XEC
-                       : FormatMoney(AVALANCHE_DEFAULT_MIN_QUORUM_STAKE));
+            fAvalanche
+                ? FormatMoney(int64_t(1'000'000'000'000) * SATOSHI) // 10B XEC
+                : FormatMoney(AVALANCHE_DEFAULT_MIN_QUORUM_STAKE));
         args.SoftSetArg(
             "-avaminquorumconnectedstakeratio",
             fAvalanche
